Zugriff fuer EnergyTraderTypeLogger Refactoring Bid:Quantity als Integer Tests fixed.

+++ src/hsoines/oekoflex/bid/Bid.java
@@ -17,7 +17,7 @@ public interface Bid {
 
     float getPrice();
 
-    float getQuantity();
+    int getQuantity();
 
     String getTypeString();
